Removed Randomizers
Removed randomized parameters from workbooks 3 and 6 -- all problems are now static. The "setup" chunk is run once to render the html document and then a second time to support exercises via the shiny server. This causes inconsistencies in problem statements and solution expectations. Considering a workaround, but problems will be static for now.